% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rconf_data.R
\name{all_agents}
\alias{all_agents}
\title{Load the agent data from a stringmol .conf file
fn is the file name
verbose whether to run quietly or not
summarize whether to compress the data into a summary of each species}
\usage{
all_agents(fn, verbose = F, getstate = F, summarize = F)
}
\description{
Load the agent data from a stringmol .conf file
fn is the file name
verbose whether to run quietly or not
summarize whether to compress the data into a summary of each species
}
