% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseSMRlist.R
\name{baseSMRlist}
\alias{baseSMRlist}
\title{Create a list of fields for a stringmol reaction}
\usage{
baseSMRlist()
}
\description{
Create a list of fields for a stringmol reaction
}
\examples{
p <- baseSMRlist()
}
