% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_type.R
\name{ccplot}
\alias{ccplot}
\title{Figure of number of mutual-replicating vs parasitic reactions}
\usage{
ccplot(rates, title, dox = F, doy = F)
}
\arguments{
\item{rates}{datframe of the two reaction rates: bind probability (bprob) and}

\item{title}{plot title}

\item{dox}{whether to plot x axis}

\item{doy}{whether to plot y axis}
}
\description{
Figure of number of mutual-replicating vs parasitic reactions
}
