% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doComplement}
\alias{doComplement}
\title{Carry out a Smith-Waterman alignment}
\usage{
doComplement(input, verbose = FALSE)
}
\arguments{
\item{input}{the input string (Assume error checking in R...)}

\item{verbose}{whether to print output or not}
}
\description{
Carry out a Smith-Waterman alignment
}
