% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doReaction}
\alias{doReaction}
\title{React 2 stringmols together - determine whether the run is deterministic or not}
\usage{
doReaction(seqVector, verbose = FALSE, climit = 1000L)
}
\arguments{
\item{seqVector}{the sequence of the two strings, active first, then passive.}
}
\description{
React 2 stringmols together - determine whether the run is deterministic or not
}
