% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doReactionFP}
\alias{doReactionFP}
\title{React 2 stringmols together - determine whether the run is deterministic or not - write to a file}
\usage{
doReactionFP(seqVector, fnVector, verbose = FALSE)
}
\arguments{
\item{seqVector}{the sequence of the two strings, active first, then passive.}
}
\description{
React 2 stringmols together - determine whether the run is deterministic or not - write to a file
}
