% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doSWAlign}
\alias{doSWAlign}
\title{Carry out a Smith-Waterman alignment}
\usage{
doSWAlign(seqVector, strip = FALSE, verbose = FALSE)
}
\arguments{
\item{seqVector}{the sequence of the two strings, active first, then passive.}
}
\description{
Carry out a Smith-Waterman alignment
}
