% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_type.R
\name{figlentime}
\alias{figlentime}
\title{Figure of molecule length and reaction time}
\usage{
figlentime(
  rundata,
  prop = 0.1,
  ylim = NULL,
  ylim2 = NULL,
  plax,
  prax,
  pbax,
  mymar = c(0.5, 0.5, 0.5, 0.5),
  cols = c("orange", "blue"),
  rno = NULL,
  plotlen = T,
  quartiles = T
)
}
\arguments{
\item{rundata}{the reaction data from the run}

\item{prop}{the proportion of all reacitons to be included (defaults to 0.1, i.e the top 10 percent)}

\item{ylim}{axis limits for the length (left) axis}

\item{ylim2}{axis limits for the reaction time (right) axis}

\item{plax}{whether to plot the left axis}

\item{prax}{whether to plot the right axis}

\item{pbax}{whether to plot the bottom axis}

\item{mymar}{the figure margins}

\item{cols}{the plot colours (defaults to orange and blue)}

\item{rno}{the run number}

\item{plotlen}{whether to plot the length}

\item{quartiles}{whether to plot the quartiles}
}
\description{
Figure of molecule length and reaction time
}
