% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_type.R
\name{figpopdy}
\alias{figpopdy}
\title{Figure of population dynamics in a stringmol run}
\usage{
figpopdy(
  rundata,
  popd,
  plax = F,
  prax = F,
  pbax = F,
  mymar = c(0.5, 0.5, 0.5, 0.5),
  cols = c("orange", "blue"),
  rno = NULL
)
}
\arguments{
\item{rundata}{the reaction data from the run}

\item{popd}{the population data from the run (including unbound molecules)}

\item{plax}{whether to plot the bottom axis}

\item{mymar}{the figure margins}

\item{cols}{the plot colours (defaults to orange and blue)}

\item{rno}{the run number}
}
\description{
Figure of population dynamics in a stringmol run
}
