% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_type.R
\name{figrepvpar}
\alias{figrepvpar}
\title{Figure of number of mutual-replicating vs parasitic reactions}
\usage{
figrepvpar(
  rundata,
  ylim = NULL,
  nsteps = F,
  plax,
  prax,
  pbax,
  mymar = c(0.5, 0.5, 0.5, 0.5),
  cols = c("orange", "blue"),
  ratio = F,
  v2 = T
)
}
\arguments{
\item{rundata}{the reaction data from the run}

\item{ylim}{y axis limits}

\item{plax}{whether to plot the left axis}

\item{prax}{whether to plot the right axis}

\item{pbax}{whether to plot the bottom axis}

\item{mymar}{the figure margins}

\item{cols}{the plot colours (defaults to orange and blue)}

\item{ratio}{whether to plot the ratio}

\item{v2}{whether to count all replicators (v2=F) or just mututal replicators (v2=T)}
}
\description{
Figure of number of mutual-replicating vs parasitic reactions
}
