% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnnhelpers.R
\name{getallspp}
\alias{getallspp}
\title{Get all the species in a run}
\usage{
getallspp(rundata, froot, maxtime = 2e+06, timestep = 20000)
}
\arguments{
\item{rundata}{data object of}

\item{froot}{the root of the filename}
}
\value{
a list of dataframes holding the following fields:
  \item{t}{the timestep}
  \item{n}{The number of individuals}
  \item{seq}{The species sequence}
}
\description{
Get all the species in a run
}
