% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splist_stats.R
\name{makeanchead}
\alias{makeanchead}
\title{get the molecules from the top 10 reactions at t=2m:}
\usage{
makeanchead(rundata, rdmanc, verbose = F)
}
\arguments{
\item{rundata}{the reaction data object for this run}

\item{rdmanc}{the ancestry data object for this run}

\item{verbose}{verbose output}
}
\value{
a data frame listing the details of the molecules in the top 10 reactions
}
\description{
get the molecules from the top 10 reactions at t=2m:
}
