% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splist_stats.R
\name{makemanc}
\alias{makemanc}
\title{Make the master ancestry data object}
\usage{
makemanc(froot)
}
\arguments{
\item{froot}{path to the "splist" data files output from a stringmol run}
}
\value{
a master ancestry data object
}
\description{
Make the master ancestry data object
}
