% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_type.R
\name{network.properties}
\alias{network.properties}
\title{Identify the network-level properties in a reaction}
\usage{
network.properties(nwp, dummy.reverse.pwp = NULL)
}
\arguments{
\item{fn}{the file name}
}
\description{
Identify the network-level properties in a reaction
}
