% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinplots.R
\name{onevio}
\alias{onevio}
\title{plot a single vioplot in analysing a run}
\usage{
onevio(
  data,
  cname = "",
  text = "",
  start = 0,
  from = 0,
  to = 2e+06,
  by = 20000,
  ptype,
  wex = 15,
  col = "black",
  doplot = F
)
}
\arguments{
\item{data}{is the datafile - an Rstringmol data frame of properties}

\item{cname}{the name of property column that is to be plotted. Set to NULL to plot all}

\item{text}{the name of the plot}

\item{start}{the lower x axis limit}

\item{from}{the first sample time}

\item{to}{the last sample time}

\item{by}{the stepsize between samples}

\item{ptype}{the plot type, can be "beta", "k", "len" and "plen", and "nos" (for raw numbers)}

\item{wex}{the scalar for the vioplot width. defaults to 15}

\item{col}{the colour of the vioplot. defaults to "black"}
}
\description{
plot a single vioplot in analysing a run
}
