% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_type.R
\name{opcol}
\alias{opcol}
\title{Obtain the species count and population sizes from a stringmol datafile
From https://www.dataanalytics.org.uk/make-transparent-colors-in-r/}
\usage{
opcol(col, percent = 70)
}
\arguments{
\item{col}{the colour}

\item{percent}{the percentage transparency}
}
\description{
Obtain the species count and population sizes from a stringmol datafile
From https://www.dataanalytics.org.uk/make-transparent-colors-in-r/
}
