% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_type.R
\name{pairwise.properties.from.conf}
\alias{pairwise.properties.from.conf}
\title{Create a data structure from a *.conf* file for passing into the standard *pairwise.properties* function}
\usage{
pairwise.properties.from.conf(fn)
}
\arguments{
\item{fn}{the file name}
}
\description{
Create a data structure from a *.conf* file for passing into the standard *pairwise.properties* function
}
