% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_type.R
\name{pairwise.properties}
\alias{pairwise.properties}
\title{Identify the network-level properties in a reaction}
\usage{
pairwise.properties(fdata, dummy.result = NULL)
}
\arguments{
\item{fdata}{data frame containing actseq and passeq fields}

\item{dummy.result}{an optional result of runReactionFP passed in (for testing)}
}
\description{
Identify the network-level properties in a reaction
}
