% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splist_stats.R
\name{plotphyl}
\alias{plotphyl}
\title{get the molecules from the top 10 reactions at t=2m:}
\usage{
plotphyl(phyl, colbylen = T, xlim = NULL)
}
\arguments{
\item{phyl}{the phylogeny}

\item{colbylen}{whether to colour the nodes by the length of the molecule}

\item{xlim}{the xlim for the plot}
}
\value{
a data frame listing the details of the molecules in the top 10 reactions
}
\description{
get the molecules from the top 10 reactions at t=2m:
}
