% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_type.R
\name{propfromseqs}
\alias{propfromseqs}
\title{Obtain the reaction properties from an active and passive sequence}
\usage{
propfromseqs(actseq, passeq)
}
\arguments{
\item{actseq}{the active sequence}

\item{passeq}{the passive sequence}
}
\description{
Obtain the reaction properties from an active and passive sequence
}
