% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnnhelpers.R
\name{qa2}
\alias{qa2}
\title{QNN analysis of a popdy object from the QNN package}
\usage{
qa2(x, scale = FALSE)
}
\arguments{
\item{x}{the popdy object}

\item{scale}{boolean, whether to divide the score by the current population size, default FALSE}
}
\value{
qnn score for the popdy object
}
\description{
QNN analysis of a popdy object from the QNN package
}
