% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rconf_data.R
\name{rconf_agents}
\alias{rconf_agents}
\title{Load the unbound agent data from a stringmol .conf file
fn is the file name
verbose whether to run quietly or not
summarize whether to compress the data into a summary of each species}
\usage{
rconf_agents(fn, verbose = F, summarize = F)
}
\description{
Load the unbound agent data from a stringmol .conf file
fn is the file name
verbose whether to run quietly or not
summarize whether to compress the data into a summary of each species
}
