% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rconf_params.R
\name{rconf_params}
\alias{rconf_params}
\title{Load the data from a stringmol .conf file
fn is the file name
verbose whether to run quietly or not
summarize whether to compress the data into a summary of each species}
\usage{
rconf_params(fn, verbose = F)
}
\description{
Load the data from a stringmol .conf file
fn is the file name
verbose whether to run quietly or not
summarize whether to compress the data into a summary of each species
}
