% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_type.R
\name{reaction_typeFP}
\alias{reaction_typeFP}
\title{Determine the type of stringmol reacion}
\usage{
reaction_typeFP(act, pas)
}
\arguments{
\item{act}{the active stringmol}

\item{pas}{the passive stringmol}
}
\description{
Determine the type of stringmol reacion
}
\examples{
p <- reaction_type("AAA","NNN")
}
\keyword{"Mass}
\keyword{spectrum",}
