% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_type.R
\name{reaction_typeFP_old}
\alias{reaction_typeFP_old}
\title{Determine the type of stringmol reaction}
\usage{
reaction_typeFP_old(act, pas, verbose = F, detail = F)
}
\arguments{
\item{act}{the active stringmol}

\item{pas}{the passive stringmol}
}
\description{
Determine the type of stringmol reaction
}
\examples{
p <- reaction_type("AAA","NNN")
}
\keyword{"Mass}
\keyword{spectrum",}
