% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_type.R
\name{runproplist}
\alias{runproplist}
\title{Obtain the reaction properties of all the reactions in a stringmol run}
\usage{
runproplist(
  froot,
  from = 20000,
  to = 2e+06,
  step = 20000,
  outfn = NULL,
  verbose = F
)
}
\arguments{
\item{froot}{the pathway to the data files}

\item{from}{the first timestep}

\item{to}{the last timestep}

\item{step}{the timestep increment}

\item{outfn}{the name of an R data object to save it to}

\item{verbose}{verbose output}
}
\description{
Obtain the reaction properties of all the reactions in a stringmol run
}
