% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runReactionFP.R
\name{runReactionFP}
\alias{runReactionFP}
\title{run a Stringmol reaction and get results via a file.}
\usage{
runReactionFP(reactants, verbose = F, remove = T)
}
\description{
run a Stringmol reaction and get results via a file.
}
