% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdi.R
\name{sdi}
\alias{sdi}
\title{Calculate Simpson's Diversity Index D1}
\usage{
sdi(pdata, time)
}
\arguments{
\item{pdata}{popdy data file loaded from a stringmol output file}

\item{time}{the time for which the index will be calculated}
}
\value{
The Simpson's Diversity Index D1
}
\description{
Calculate Simpson's Diversity Index D1
}
