% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqstats.R
\name{seqstats}
\alias{seqstats}
\title{Calculate Stats on sequence data:}
\usage{
seqstats(pdata, time = NA)
}
\arguments{
\item{pdata}{popdy data file loaded from a stringmol output file}

\item{time}{the time for which the index will be calculated}
}
\value{
The Simpson's Diversity Index D1
}
\description{
Calculate Stats on sequence data:
}
