% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_type.R
\name{spdata}
\alias{spdata}
\title{Obtain the species count and population sizes from a stringmol datafile}
\usage{
spdata(froot = "", trange = seq(20000, 2e+06, 20000))
}
\arguments{
\item{froot}{the path and file stem of the data files, e.g. "~/Desktop/smsp/out3/out1_%d.conf"}

\item{trange}{the time range and step size of the log files}
}
\description{
Obtain the species count and population sizes from a stringmol datafile
}
