% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splist_makemaster.R
\name{splist_makemaster}
\alias{splist_makemaster}
\title{Try to make a master list of all 'seen' species, handling issues with stringmol logging over restarts.}
\usage{
splist_makemaster(fn, tmin = 0, verbose = F)
}
\arguments{
\item{fn}{the file name of the species list}

\item{tmin}{the start time}

\item{verbose}{messages or not}
}
\value{
species list
}
\description{
Try to make a master list of all 'seen' species, handling issues with stringmol logging over restarts.
}
