% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splist_stats.R
\name{splist_stats}
\alias{splist_stats}
\title{Load a splist file and format the data}
\usage{
splist_stats(fn, tmin = 0, verbose = F)
}
\arguments{
\item{fn}{the file name of the species list}

\item{tmin}{the start time}

\item{verbose}{messages or not}
}
\value{
species list
}
\description{
Load a splist file and format the data
}
