% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnnhelpers.R
\name{unboundpopfromconf}
\alias{unboundpopfromconf}
\title{Get the unbound molecular species counts from a stringmol config file}
\usage{
unboundpopfromconf(froot, tt, verbose = F)
}
\arguments{
\item{froot}{the root of the filename}

\item{tt}{the timestep of the config file}

\item{verbose}{whether to write messages whilst processing}
}
\value{
a dataframe holding the following fields:
  \item{t}{the timestep}
  \item{n}{The number of individuals}
  \item{seq}{The species sequence}
}
\description{
Get the unbound molecular species counts from a stringmol config file
}
