% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_type.R
\name{well.mixed.dynamics}
\alias{well.mixed.dynamics}
\title{Deterministic calculation of the carrying capacity of a self-replicator}
\usage{
well.mixed.dynamics(
  bprob,
  nsteps,
  initpop = 100,
  maxt = 20000,
  death = 5e-04,
  lim = (125 * 100)
)
}
\arguments{
\item{bprob}{the bind probability}

\item{nsteps}{the number of timesteps to replicate}

\item{initpop}{the initial population}

\item{maxt}{the time limit on the calculation}

\item{death}{the death rate}

\item{lim}{the arena size}
}
\description{
Deterministic calculation of the carrying capacity of a self-replicator
}
